pyproject.toml
[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"

[tool.poetry]
name = "fast-api-service"
version = "0.1.0"
description = "Fast API service"
authors = ["Keyur Parekh"]

[tool.poetry.dependencies]
python = "^3.12"
fastapi = "^0.115.6"
uvicorn = "^0.32.1"
python-dotenv = "^1.0.1"
prometheus-fastapi-instrumentator = "^7.0.0"
structlog = "^24.4.0"
asgi-correlation-id = "^4.3.3"
logging-http-client = "^2.32.3.5"
google-cloud-secret-manager = "^2.16.4"
google-cloud-storage = "^2.10.0"
xmltodict = "^0.14.2"
pydantic-xml = "^2.14.1"

[tool.poetry.group.dev.dependencies]
black = "^24.2.0"
flake8 = "^7.1.0"
flake8-pyproject = "^1.2.3"
pyclean = "^3.0.0"
rich = "^13.7.1"

[tool.poetry.group.test.dependencies]
pytest = "^8.1.1"
pytest-mock = "^3.14.0"
pytest-env = "^1.1.3"
pytest-order = "^1.3.0"
testcontainers = "^4.7.1"
wiremock= "^2.6.1"
httpx = "^0.27.0"

[tool.black]
line-length = 120

[tool.flake8]
max-line-length = 120
count = true

[tool.pytest.ini_options]
pythonpath = "src tests"
testpaths = "tests"
env = [
    "LOG_LEVEL=DEBUG",
]
